# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Medarx Application API (Phase 1 Privacy Kernel)
  version: 0.1.0-phase1
  summary: HTTP surface for the Medarx Phase 1 privacy kernel.
  description: |
    Medarx is intended for research and demonstration. It performs administrative report assistance using clinician-supplied findings and authorized report text. It does not independently interpret medical images, make diagnostic or treatment recommendations, or autonomously submit clinical reports. Its regulatory classification has not been established, and it is not validated for clinical use.

    This document specifies component J (Application API) of the approved Phase 1 design: the HTTP surface that exercises components A through I. It is the privacy kernel's boundary, not a clinical product surface.

    ## Relationship to the design document

    The normative source for enforcement behaviour is §6 (Fail-Closed Enforcement) of the design. This specification is a faithful encoding of that table; the `layer` values used in the `BlockReceipt` correspond one-to-one with the design's layer designations (`A`, `D.1`, `D.2`, `D.3`, `E`, `F`).

    ## Two things this API deliberately does NOT do

...

...

    ## Function scope

...

    ## Status-code contract

    Two refusal conditions must be distinguishable **by status code alone**, without parsing the body:

    | Status | Meaning | Raised by |
    |---|---|---|
...

...

    ## Request identity

...

    ## Authentication

...

    ## Audit readback

...
  contact:
...
servers:
...
security: []
paths:
  /v1/functions/{function_name}/executions:
    post:
      tags: [execution]
      operationId: createFunctionExecution
      summary: Execute a named function over clinician-supplied report text and allowlisted DICOM metadata.
      description: |
        Submits an execution request for one of the three closed functions (`Draft`, `Prior Summary`,
        `Ask`) against a study context, with clinician-supplied report text and an allowlisted set of
        DICOM metadata fields, plus any prior-study references.

        On success the response is the model response **as a draft requiring explicit human sign-off**,
        together with the request ID, the selected model, the de-identification policy version, and the
        hash of the approved payload that was transmitted. No endpoint in this API submits that draft
        anywhere; recording a human decision is `recordHumanApproval`.

        On any privacy-policy block the response is `422` with a `BlockReceipt`. Raw PHI values never
        appear in a block receipt, and nothing is transmitted to a provider — in either implemented
        policy mode.

        This operation does NOT accept arbitrary DICOM objects or free-form prompts. The only DICOM
        fields expressible are those in `AllowlistedDicomMetadata`, and the only free text is
        `report_text`, which is a documented pipeline input rather than a prompt. Submitting anything
        outside those shapes is rejected at this surface and never enters the pipeline.
      parameters:
        - $ref: '#/components/parameters/FunctionNamePathParam'
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        description: The execution request.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
      responses:
        '200':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          description: >-
            The privacy pipeline approved the payload and the model gateway returned a response.
            The response is a draft; it has not been approved by a human and has not been written
            to any clinical system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/MalformedRequest'
        '403':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/Forbidden'
        '404':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/FunctionNotFound'
        '415':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/Blocked'
        '500':
          headers:
            $ref: '#/components/headers/RequestIdHeader'
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes: {}
  parameters:
    FunctionNamePathParam:
      name: function_name
      in: path
      required: true
      description: >-
        The function to execute. A closed enum: unknown functions are rejected rather than processed.
        Phase 1 implements the privacy pipeline around these functions, not their full clinical
        behaviour.
      schema:
        $ref: '#/components/schemas/FunctionName'
...
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: >-
        Optional caller-supplied request ID, honoured if present and recorded in the audit log and
        echoed in the response. If omitted, the server generates one. This is a **header**, not a
        body field: the request ID identifies the request rather than being part of the extraction
        input, and a header is still available when the body is unparseable. The design does not fix
        this choice; the spec does.
      schema:
        $ref: '#/components/schemas/RequestId'
      example: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
...
  headers:
    RequestIdHeader:
      description: >-
        The request ID of this exchange, echoed on **every** response including `400`, `403`, `415`
        and `422`. It is the caller's `X-Request-Id` when one was supplied and a server-generated ID
        when one was not, it is the same value the audit record is filed under, and it is present
        even on a response whose body is unparseable — which is the whole reason it is a header and
        not a body field. Declared here rather than left to each response, so the promise is written
        down once.
      schema:
        type: string
  responses:
    Blocked:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        Privacy-policy block. The request was well-formed and authorized, but the privacy pipeline
        could not produce a policy-compliant payload, so **nothing was transmitted** to a provider in
        either implemented mode. Raw PHI values never appear in this body.

        This is the `422` counterpart to `403`. A caller distinguishes a privacy block from an
        authorization failure by status code alone: `422` here, `403` there.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BlockReceipt'
          examples:
            unresolvedNerCandidate:
              summary: Design §6 row 4 — layer 2 NER candidate below confidence threshold, or with no deterministic replacer.
              value:
                status: blocked
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                layer: D.2
                action_codes: [NER_UNRESOLVED]
                policy_version: ps315-baseline-v1
            fieldOutsideAllowlist:
              summary: Design §6 row 2 — field outside the function allowlist; rejected by A, no downstream processing.
              value:
                status: blocked
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                layer: A
                action_codes: [FIELD_NOT_ALLOWLISTED, UNKNOWN_FUNCTION]
                policy_version: ps315-baseline-v1
            deterministicReplacementFailed:
              summary: Design §6 row 3 — layer 1 could not deterministically remove or replace a required structured field.
              value:
                status: blocked
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                layer: D.1
                action_codes: [DETERMINISTIC_REPLACEMENT_FAILED]
                policy_version: ps315-baseline-v1
            contractViolation:
              summary: >-
                Design §6 row 5 — layer 3 contract violation (missing/extra field, unshifted date, missing
                surrogate, leftover pattern match, or hash mismatch).
              value:
                status: blocked
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                layer: D.3
                action_codes: [UNSHIFTED_DATE, LEFTOVER_PATTERN_MATCH]
                policy_version: ps315-baseline-v1
            policyEngineFailClosed:
              summary: >-
                Design §6 row 6 — unresolved disposition, unknown or missing policy version, or policy
                configuration error. Blocks in every implemented mode.
              value:
                status: blocked
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                layer: E
                action_codes: [UNRESOLVED_DISPOSITION, UNKNOWN_POLICY_VERSION, POLICY_CONFIG_ERROR]
                policy_version: ps315-baseline-v1
            gatewayRejection:
              summary: >-
                Design §6 row 7 — payload was not the exact object E approved, or the model identifier is
                unknown. No provider call is made.
              value:
                status: blocked
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                layer: F
                action_codes: [PAYLOAD_MISMATCH, UNKNOWN_MODEL]
                policy_version: ps315-baseline-v1
    MalformedRequest:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        The request could not be understood as a request at all: unparseable body, wrong media type,
        or a JSON document that does not match the request schema. This is a request-construction
        failure (design §6 row 1) and is distinct from both `403` (authorization) and `422`
        (privacy block). Nothing entered the pipeline.

        Where the request is well-formed JSON but violates the schema, the body is a `ValidationError`
        listing the offending fields. Where the body is unparseable, it is a `ProblemDetail`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            schemaViolation:
              summary: Well-formed JSON that does not match the execution request schema.
              value:
                type: https://medarx.invalid/problems/request-validation
                title: Request does not match the execution request schema
                status: 400
                detail: >-
                  Unrecognized or disallowed properties present: dicom_object, prompt. Medarx does not
                  accept arbitrary DICOM objects or free-form prompts at this surface.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
                errors:
                  - field: body.dicom_object
                    message: Additional properties are not permitted.
                  - field: body.prompt
                    message: Additional properties are not permitted.
            unparseableBody:
              summary: The body could not be parsed as JSON.
              value:
                type: https://medarx.invalid/problems/malformed-request
                title: Malformed request
                status: 400
                detail: Request body is not valid JSON.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
    Forbidden:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        Authorization failure: the caller is not permitted to execute this function, or the study
        context is outside the caller's authorized scope (design §6 row 1).

        This is `403` and never `422`. It is deliberately distinct from a privacy-policy block: the
        request here was not permitted in the first place, whereas a `422` means the request was
        permitted and well-formed but the privacy pipeline refused it. A caller must be able to tell
        these apart by status code alone, without parsing the body.

        Phase 1 exercises a **synthetic** scope; the design's per-user authorization model (roles,
        per-user permissions) is deferred to Phase 6. No authentication scheme is defined for this
        API.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            functionNotPermitted:
              summary: The function is not permitted for this caller.
              value:
                type: https://medarx.invalid/problems/function-not-permitted
                title: Function not permitted
                status: 403
                detail: The requested function is not permitted in the caller's scope.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
            studyScopeUnauthorized:
              summary: The study context is outside the caller's authorized scope.
              value:
                type: https://medarx.invalid/problems/study-scope-unauthorized
                title: Study scope not authorized
                status: 403
                detail: The requested study is not within the caller's authorized scope.
                request_id: 7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34
    UnsupportedMediaType:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        The request media type is not `application/json`. This API accepts structured JSON only; it
        does not accept DICOM multipart payloads, so `multipart/related` and `application/dicom` are
        rejected here by design.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    FunctionNotFound:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        The `function_name` path segment does not name one of the three closed functions
        (`Draft`, `Prior Summary`, `Ask`). The function set is closed, not free-form, so an
        unrecognized name is not processed; a request naming an unknown function never enters the
        pipeline. This path-shaped case is reported as `404` because no such function resource
        exists; the same condition arriving as a body field against a valid path is reported as a
        `422` block with `layer: A` and `action_codes` containing `UNKNOWN_FUNCTION`, since the design
        assigns unknown-function rejection to component A.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    UnexpectedError:
      headers:
        $ref: '#/components/headers/RequestIdHeader'
      description: >-
        An unexpected server-side failure. Sensitive-data filtering is applied to application logs,
        exception traces, and tracing spans under the same policy as the audit log, so no raw PHI is
        disclosed through this response or its logs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    FunctionName:
      type: string
      title: FunctionName
      description: >-
        The closed set of Medarx functions. The set is closed by design: functions are a fixed,
        enumerated surface, not free-form.

        Phase 1 implements the **privacy pipeline around** these functions, not their full clinical
        behaviour. Phase 1 exercises the pipeline with report text and metadata; `Draft`'s automated
        output validation is a Phase 4 concern, and `Prior Summary` / `Ask` contract validation is a
        Phase 5 concern. What Phase 1 proves for each is authorization, allowlisted extraction,
        pseudonymization, layered redaction, and policy approval or refusal before transmission.
      enum:
        - Draft
        - Prior Summary
        - Ask
    RequestId:
      type: string
      description: >-
        Identifier of a single API request. Generated by the server when the caller does not supply
        one; carried on every block receipt, every successful execution response, and every audit
        record.
      examples: [7f3c1a90-2b6e-4f1a-9d55-2c8b7e0a1f34]
...
    PolicyMode:
      type: string
      description: >-
        The configured policy mode (component E). Fixed per deployment; never selectable per request.
      enum:
        - strict_local
        - cloud
        - authorized_local
...
    StudyContext:
      type: object
      additionalProperties: false
      required:
        - study_reference
      description: >-
        Study-scoped context for the request. In Phase 1 all study and prior records are synthetic.
        The study reference is an internal reference; the pseudonymization service exchanges it for a
        stable surrogate before anything is sent to a model provider.
      properties:
        study_reference:
          type: string
          description: >-
            Internal study reference. Replaced by a stable surrogate identifier within an authorized
            scope before the payload reaches the model gateway. It is not a DICOM StudyInstanceUID and
            is never transmitted as-is.
          examples: ['STUDY-SYN-000041']
        patient_reference:
          type: string
          description: >-
            Internal patient reference, if the caller knows it. This is the reference the
            de-identification policy's **per-patient** date shift is chosen by, and supplying it is
            what makes design §3 C's stated invariant — "the date shift preserves sequence and
            duration" — hold *across* two studies of one patient rather than only within one study.

            Optional, and a request that omits it is not refused: the shift still applies, and the
            fallback is the `PatientID` in `dicom_metadata` where the caller supplied one, and the
            study reference otherwise. **The study-reference fallback is a weaker guarantee** — a
            study is not a patient, so two studies of one patient that omit both references are
            shifted by two unrelated offsets, the interval between them is not preserved, and can
            invert. The reference is a pseudonymization key: it is never carried in a payload, never
            reaches a model provider, and is not stored in the audit log, whose storage policy has no
            field for a subject reference and must not gain one.
          examples: ['PAT-SYN-000041']
        accession_reference:
          type: string
          description: >-
            Internal accession reference, if known. Detected by the mandatory custom clinical
            recognizer for accession number; surrogate-substituted on the same path.
          examples: ['ACC-SYN-0000417']
        encounter_reference:
          type: string
          description: Internal encounter reference, if known. Surrogate-substituted on the same path.
          examples: ['ENC-SYN-0000417']
        modality:
          type: string
          description: >-
            Modality of the source study, as reported. Metadata only — Medarx performs no diagnostic
            image interpretation and never derives findings from pixels.
          examples: [CT]
...
    ReportText:
      type: object
      additionalProperties: false
      required:
        - text
      description: >-
        Clinician-supplied report text — the single free-text input on this surface.

        It is a documented pipeline input, **not** a prompt. It is not a free-form instruction to a
        model: it is allowlisted content that the extraction, pseudonymization, and redaction layers
        process before any model call. Free-form prompts are rejected at this surface.

        Text is subject to the full redaction pipeline: a candidate in this field that the detectors
        cannot confidently sanitize is exactly the load-bearing block case (design §4 Beat 2).
      properties:
        text:
          type: string
          minLength: 1
          description: >-
            The report text as the clinician wrote it, before pseudonymization and redaction. It is
            never transmitted to a model provider in this form.
          examples:
            - 'FINDINGS: 7 mm nodule in the right upper lobe. Correlate with prior MRN 4452819.'
        source:
          type: string
          description: Where the text came from, for the audit record's input-source attribution.
          enum:
            - synthetic_corpus
            - clinician_supplied
    ExecutionRequest:
      type: object
      additionalProperties: false
      required:
        - study_context
        - report_text
      description: >-
        An execution request. Every property is one the design places on this surface: a function, a
        study context, report text, an allowlisted set of DICOM metadata fields, and any prior-study
        references.

        The schema closes on `additionalProperties: false` precisely so that an arbitrary DICOM
        object, a raw pixel/element payload, or a free-form prompt field cannot be supplied. Unknown
        properties are a malformed request (`400`), and any attempt to smuggle one in through a
        property like `dicom_object`, `pixel_data`, or `prompt` is rejected at the surface with
        nothing entering the pipeline.

        There is deliberately **no** `policy_mode` property: the policy mode is a deployment
        configuration, not a caller choice.
      properties:
        function:
          $ref: '#/components/schemas/FunctionName'
        study_context:
          $ref: '#/components/schemas/StudyContext'
        report_text:
          $ref: '#/components/schemas/ReportText'
        dicom_metadata:
          $ref: '#/components/schemas/AllowlistedDicomMetadata'
        prior_studies:
          type: array
          description: Prior-study references relevant to this request, if any.
          items:
            $ref: '#/components/schemas/PriorStudyReference'
        model_id:
          type: string
          description: >-
            Optional model identifier for component F, validated against the model registry. An
            unknown identifier causes a fail-closed block attributed to `F` with no provider call
            made. Omit to use the deployment's configured default.
          examples: ['openrouter/default']
        requested_language:
          type: string
          description: Optional output language hint passed through to the gateway. Not a prompt.
          examples: [en]
...
    ExecutionResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - request_id
        - draft
        - policy_version
      description: >-
        The approved-path response: the model response plus the request ID, the identity of the
        approved payload that was transmitted, and the policy version under which it was approved.

        The payload is identified by hash. The hash is what egress verification checks observed
        outbound bytes against, using two observers that are both external to the Medarx process —
        Medarx's own audit log is explicitly not accepted as evidence of what left.
      properties:
        status:
          type: string
          const: approved
          description: The pipeline approved the payload and the gateway returned a response.
        request_id:
          $ref: '#/components/schemas/RequestId'
        function:
          $ref: '#/components/schemas/FunctionName'
        policy_version:
          $ref: '#/components/schemas/PolicyVersion'
        policy_mode:
          $ref: '#/components/schemas/PolicyMode'
        selected_model:
          type: string
          description: The model the gateway used.
        approved_payload_hash:
          type: string
          description: >-
            Hash of the exact approved payload object transmitted. Computed at redaction layer 3 and
            verified against observed egress bytes by component I.
          examples: ['sha256:02f1a0c9d4b7e6a3f5c8d1e0b2a4f6c8d0e2b4a6c8e0f2d4b6a8c0e2f4d6b8a0']
        input_hash:
          type: string
          description: Hash of the caller's input as received, for audit correlation.
        draft:
          allOf:
            - $ref: '#/components/schemas/ModelResponse'
          description: >-
            The model response, presented as a draft requiring explicit human sign-off. It has **not**
            been approved and has not been written to any clinical system.
...